Code refactoring for Cards and NewQuestCard
